+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Swapped from `proc-macro-error2` to `syn::Error`
 - Renamed the API for getting the length of an inner block of registers (from
   `XXX_array_len` to `len_XXX`)
 - The generated wrapper struct now `#[repr(transparent)]`.